Room Changing
The room will now change when the player is not looking at them. This has some issues for now like changing the tile the player is standing on, the player can also just look up. All tiles change instead of just the ones near the player. The final room will not despawn because it is locked.